requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
python-dotenv==1.2.1
markdown==3.6
groq==0.8.0
MarkupSafe==3.0.2
itsdangerous==2.2.0
Werkzeug==3.1.3
Jinja2==3.1.6